+++ release-notes.rst
@@ -2,6 +2,24 @@ Release notes for chevah.compat
 ===============================
 
 
+0.71.2 - 2023-05-22
+-------------------
+
+* Fix Unicode path returned on Windows for unlocked avatars.
+
+
+0.71.1 - 2023-05-22
+-------------------
+
+* Fix getAbsoluteReal path Windows not to return a path in long path format.
+
+
+0.71.0 - 2023-05-20
+-------------------
+
+* Add support for long paths in Windows.
+
+
 0.70.3 - 2023-04-18
 -------------------